+++ common/types/activation.go
@@ -2,7 +2,6 @@ package types
 
 import (
 	"encoding/hex"
-	"errors"
 	"time"
 
 	"github.com/spacemeshos/go-scale"
@@ -163,22 +162,22 @@ type ActivationTx struct {
 	PrevATXID ATXID
 
 	// CommitmentATX is the ATX used in the commitment for initializing the PoST of the node.
-	CommitmentATX *ATXID
-	Coinbase      Address
-	NumUnits      uint32
-
-	VRFNonce *VRFPostIndex
+	CommitmentATX  *ATXID
+	Coinbase       Address
+	NumUnits       uint32 // the minimum number of space units in this and the previous ATX
+	BaseTickHeight uint64
+	TickCount      uint64
+	VRFNonce       *VRFPostIndex
 
 	SmesherID NodeID
 	Signature EdSignature
 
 	AtxBlob
 
-	golden            bool
-	id                ATXID     // non-exported cache of the ATXID
-	effectiveNumUnits uint32    // the number of effective units in the ATX (minimum of this ATX and the previous ATX)
-	received          time.Time // time received by node, gossiped or synced
-	validity          Validity  // whether the chain is fully verified and OK
+	golden   bool
+	id       ATXID     // non-exported cache of the ATXID
+	received time.Time // time received by node, gossiped or synced
+	validity Validity  // whether the chain is fully verified and OK
 }
 
 // NewActivationTx returns a new activation transaction. The ATXID is calculated and cached.
@@ -217,6 +216,36 @@ func (atx *ActivationTx) SetGolden() {
 	atx.golden = true
 }
 
+// Weight of the ATX. The total weight of the epoch is expected to fit in a uint64 and is
+// sum(atx.NumUnits * atx.TickCount for each ATX in a given epoch).
+// Space Units sizes are chosen such that NumUnits for all ATXs in an epoch is expected to be < 10^6.
+// PoETs should produce ~10k ticks at genesis, but are expected due to technological advances
+// to produce more over time. A uint64 should be large enough to hold the total weight of an epoch,
+// for at least the first few years.
+func (atx *ActivationTx) GetWeight() uint64 {
+	return getWeight(uint64(atx.NumUnits), atx.TickCount)
+}
+
+// TickHeight returns a sum of base tick height and tick count.
+func (atx *ActivationTx) TickHeight() uint64 {
+	return atx.BaseTickHeight + atx.TickCount
+}
+
+func (atx *ActivationTx) ToHeader() *ActivationTxHeader {
+	return &ActivationTxHeader{
+		PublishEpoch:      atx.PublishEpoch,
+		Coinbase:          atx.Coinbase,
+		EffectiveNumUnits: atx.NumUnits,
+		Received:          atx.Received(),
+
+		ID:     atx.ID(),
+		NodeID: atx.SmesherID,
+
+		BaseTickHeight: atx.BaseTickHeight,
+		TickCount:      atx.TickCount,
+	}
+}
+
 // MarshalLogObject implements logging interface.
 func (atx *ActivationTx) MarshalLogObject(encoder log.ObjectEncoder) error {
 	encoder.AddString("atx_id", atx.id.String())
@@ -233,10 +262,11 @@ func (atx *ActivationTx) MarshalLogObject(encoder log.ObjectEncoder) error {
 	encoder.AddString("coinbase", atx.Coinbase.String())
 	encoder.AddUint32("epoch", atx.PublishEpoch.Uint32())
 	encoder.AddUint64("num_units", uint64(atx.NumUnits))
-	if atx.effectiveNumUnits != 0 {
-		encoder.AddUint64("effective_num_units", uint64(atx.effectiveNumUnits))
-	}
 	encoder.AddUint64("sequence_number", atx.Sequence)
+	encoder.AddUint64("base_tick_height", atx.BaseTickHeight)
+	encoder.AddUint64("tick_count", atx.TickCount)
+	encoder.AddUint64("weight", atx.GetWeight())
+	encoder.AddUint64("height", atx.TickHeight())
 	return nil
 }
 
@@ -250,20 +280,9 @@ func (atx *ActivationTx) ID() ATXID {
 	return atx.id
 }
 
-func (atx *ActivationTx) EffectiveNumUnits() uint32 {
-	if atx.effectiveNumUnits == 0 {
-		panic("effectiveNumUnits field must be set")
-	}
-	return atx.effectiveNumUnits
-}
-
 // SetID sets the ATXID in this ATX's cache.
 func (atx *ActivationTx) SetID(id ATXID) {
 	atx.id = id
-}
-
-func (atx *ActivationTx) SetEffectiveNumUnits(numUnits uint32) {
-	atx.effectiveNumUnits = numUnits
 }
 
 func (atx *ActivationTx) SetReceived(received time.Time) {
@@ -280,23 +299,6 @@ func (atx *ActivationTx) Validity() Validity {
 
 func (atx *ActivationTx) SetValidity(validity Validity) {
 	atx.validity = validity
-}
-
-// Verify an ATX for a given base TickHeight and TickCount.
-func (atx *ActivationTx) Verify(baseTickHeight, tickCount uint64) (*VerifiedActivationTx, error) {
-	if atx.effectiveNumUnits == 0 {
-		return nil, errors.New("effective num units not set")
-	}
-	if !atx.Golden() && atx.received.IsZero() {
-		return nil, errors.New("received time not set")
-	}
-	vAtx := &VerifiedActivationTx{
-		ActivationTx: atx,
-
-		baseTickHeight: baseTickHeight,
-		tickCount:      tickCount,
-	}
-	return vAtx, nil
 }
 
 // Merkle proof proving that a given leaf is included in the root of merkle tree.
@@ -383,7 +385,7 @@ func ATXIDsToHashes(ids []ATXID) []Hash32 {
 
 type EpochActiveSet struct {
 	Epoch EpochID
-	Set   []ATXID `scale:"max=3500000"` // to be in line with `EpochData` in fetch/wire_types.go
+	Set   []ATXID `scale:"max=4500000"` // to be in line with `EpochData` in fetch/wire_types.go
 }
 
 var MaxEpochActiveSetSize = scale.MustGetMaxElements[EpochActiveSet]("Set")