+++ fetch/p2p_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/spacemeshos/go-spacemesh/activation/wire"
 	"github.com/spacemeshos/go-spacemesh/codec"
 	"github.com/spacemeshos/go-spacemesh/common/types"
 	"github.com/spacemeshos/go-spacemesh/datastore"
@@ -354,7 +353,7 @@ func TestP2PGetATXs(t *testing.T) {
 						context.Background(), []types.ATXID{atx.ID()})
 				},
 				errStr, "atx", "hs/1", types.Hash32(atx.ID()), atx.ID().Bytes(),
-				codec.MustEncode(wire.ActivationTxToWireV1(atx.ActivationTx)))
+				atx.AtxBlob.Blob)
 		})
 }