+++ malfeasance/handler_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
 
-	"github.com/spacemeshos/go-spacemesh/activation"
 	awire "github.com/spacemeshos/go-spacemesh/activation/wire"
 	"github.com/spacemeshos/go-spacemesh/codec"
 	"github.com/spacemeshos/go-spacemesh/common/types"
@@ -38,12 +37,10 @@ func createIdentity(t *testing.T, db *sql.Database, sig *signing.EdSigner) {
 		PublishEpoch: types.EpochID(1),
 	}
 	atx := types.NewActivationTx(challenge, types.Address{}, 1, nil)
-	require.NoError(t, activation.SignAndFinalizeAtx(sig, atx))
-	atx.SetEffectiveNumUnits(atx.NumUnits)
+	atx.SmesherID = sig.NodeID()
 	atx.SetReceived(time.Now())
-	vAtx, err := atx.Verify(0, 1)
-	require.NoError(t, err)
-	require.NoError(t, atxs.Add(db, vAtx))
+	atx.TickCount = 1
+	require.NoError(t, atxs.Add(db, atx))
 }
 
 func TestHandler_HandleMalfeasanceProof_multipleATXs(t *testing.T) {